Disable node discovery by default
Discovery of new nodes requires explicit support from BMO, which
does not seem to exist, nor is on the roadmap. Having discovery
enabled may cause hard to debug situations when an accidentally
booted IPA results in a new node added.

An option is left to enable discovery for testing purposes or
for future implementation on the BMO side.